+++ 1.20.4/Common/src/main/java/fuzs/resourcepackoverrides/client/handler/PackActionsHandler.java
@@ -12,12 +12,13 @@ import net.minecraft.client.MouseHandler;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.components.toasts.TutorialToast;
-import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.gui.screens.packs.PackSelectionScreen;
 import net.minecraft.client.gui.screens.packs.TransferableSelectionList;
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.Mth;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Collections;
 import java.util.Optional;
 
 public class PackActionsHandler {
@@ -29,12 +30,12 @@ public class PackActionsHandler {
         PACK_ACTIONS.put(InputConstants.KEY_C, new PackAction(Component.translatable("packAction.copyId.title"), Component.translatable("packAction.copyId.description", Component.literal("C").withStyle(ChatFormatting.BOLD)), Component.translatable("packAction.copyId.success")) {
 
             @Override
-            boolean execute(Minecraft minecraft) {
+            boolean execute(Minecraft minecraft, PackSelectionScreen screen) {
                 MouseHandler mouseHandler = minecraft.mouseHandler;
                 Window window = minecraft.getWindow();
                 int mouseX = (int) (mouseHandler.xpos() * window.getGuiScaledWidth() / window.getScreenWidth());
                 int mouseY = (int) (mouseHandler.ypos() * window.getGuiScaledHeight() / window.getScreenHeight());
-                Optional<String> hoveredPackId = getHoveredPackId(minecraft.screen, mouseX, mouseY);
+                Optional<String> hoveredPackId = getHoveredPackId(screen, mouseX, mouseY);
                 hoveredPackId.ifPresent(minecraft.keyboardHandler::setClipboard);
                 return hoveredPackId.isPresent();
             }
@@ -42,7 +43,7 @@ public class PackActionsHandler {
         PACK_ACTIONS.put(InputConstants.KEY_D, new PackAction(Component.translatable("packAction.toggleDebug.title"), Component.translatable("packAction.toggleDebug.description", Component.literal("D").withStyle(ChatFormatting.BOLD)), Component.translatable("packAction.toggleDebug.success")) {
 
             @Override
-            boolean execute(Minecraft minecraft) {
+            boolean execute(Minecraft minecraft, PackSelectionScreen screen) {
                 debugTooltips = !debugTooltips;
                 return true;
             }
@@ -50,25 +51,35 @@ public class PackActionsHandler {
         PACK_ACTIONS.put(InputConstants.KEY_R, new PackAction(Component.translatable("packAction.reloadSettings.title"), Component.translatable("packAction.reloadSettings.description", Component.literal("R").withStyle(ChatFormatting.BOLD)), Component.translatable("packAction.reloadSettings.success")) {
 
             @Override
-            boolean execute(Minecraft minecraft) {
+            boolean execute(Minecraft minecraft, PackSelectionScreen screen) {
                 ResourceOverridesManager.load();
-                if (minecraft.screen != null) {
-                    minecraft.screen.init(minecraft, minecraft.getWindow().getGuiScaledWidth(), minecraft.getWindow().getGuiScaledHeight());
-                }
+                screen.init(minecraft, minecraft.getWindow().getGuiScaledWidth(), minecraft.getWindow().getGuiScaledHeight());
+                return true;
+            }
+        });
+        PACK_ACTIONS.put(InputConstants.KEY_T, new PackAction(Component.translatable("packAction.restoreDefaults.title"), Component.translatable("packAction.restoreDefaults.description", Component.literal("T").withStyle(ChatFormatting.BOLD)), Component.translatable("packAction.restoreDefaults.success")) {
+
+            @Override
+            boolean execute(Minecraft minecraft, PackSelectionScreen screen) {
+                minecraft.getResourcePackRepository().setSelected(ResourceOverridesManager.getDefaultResourcePacks(true));
+                screen.model.selected.clear();
+                screen.model.selected.addAll(minecraft.getResourcePackRepository().getSelectedPacks());
+                Collections.reverse(screen.model.selected);
+                screen.init(minecraft, minecraft.getWindow().getGuiScaledWidth(), minecraft.getWindow().getGuiScaledHeight());
                 return true;
             }
         });
     }
 
-    public static void onScreen$Render$Post(Screen screen, GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
-        if (debugTooltips) {
+    public static void onScreen$Render$Post(Minecraft minecraft, PackSelectionScreen screen, GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
+        if (debugTooltips && screen.model.repository == minecraft.getResourcePackRepository()) {
             getHoveredPackId(screen, mouseX, mouseY).map(Component::literal).ifPresent(component -> {
                 guiGraphics.renderTooltip(ClientAbstractions.getScreenFont(screen), component, mouseX, mouseY);
             });
         }
     }
 
-    private static Optional<String> getHoveredPackId(@Nullable Screen screen, int mouseX, int mouseY) {
+    private static Optional<String> getHoveredPackId(PackSelectionScreen screen, int mouseX, int mouseY) {
         if (screen == null) return Optional.empty();
         for (GuiEventListener guiEventListener : screen.children()) {
             if (guiEventListener instanceof TransferableSelectionList selectionList) {
@@ -91,9 +102,11 @@ public class PackActionsHandler {
         PACK_ACTIONS.values().forEach(action -> action.tick(minecraft));
     }
 
-    public static void onKeyPressed$Post(Screen screen, int keyCode, int scanCode, int modifiers) {
-        PackAction packAction = PACK_ACTIONS.get(keyCode);
-        if (packAction != null) packAction.update();
+    public static void onKeyPressed$Post(Minecraft minecraft, PackSelectionScreen screen, int keyCode, int scanCode, int modifiers) {
+        if (screen.model.repository == minecraft.getResourcePackRepository()) {
+            PackAction packAction = PACK_ACTIONS.get(keyCode);
+            if (packAction != null) packAction.update();
+        }
     }
 
     private static abstract class PackAction {
@@ -134,8 +147,8 @@ public class PackActionsHandler {
                 }
                 if (this.pressTime < 20) {
                     this.toast.updateProgress(Mth.clamp(this.pressTime / 20.0F, 0.0F, 1.0F));
-                } else if (!this.wasExecuted) {
-                    if (this.execute(minecraft)) {
+                } else if (!this.wasExecuted && minecraft.screen instanceof PackSelectionScreen screen && screen.model.repository == minecraft.getResourcePackRepository()) {
+                    if (this.execute(minecraft, screen)) {
                         this.finish(minecraft);
                     }
                     this.wasExecuted = true;
@@ -162,7 +175,7 @@ public class PackActionsHandler {
             this.wasExecuted = false;
         }
 
-        abstract boolean execute(Minecraft minecraft);
+        abstract boolean execute(Minecraft minecraft, PackSelectionScreen screen);
 
         private void finish(Minecraft minecraft) {
             if (this.successToast != null) this.successToast.hide();