+++ 1.20.4/Fabric/src/main/java/fuzs/resourcepackoverrides/fabric/client/ResourcePackOverridesFabricClient.java
@@ -5,6 +5,9 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenKeyboardEvents;
+import net.fabricmc.fabric.api.client.screen.v1.Screens;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.packs.PackSelectionScreen;
 
 public class ResourcePackOverridesFabricClient implements ClientModInitializer {
@@ -16,9 +19,13 @@ public class ResourcePackOverridesFabricClient implements ClientModInitializer {
 
     private static void registerHandlers() {
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
-            if (screen instanceof PackSelectionScreen) {
-                ScreenEvents.afterRender(screen).register(PackActionsHandler::onScreen$Render$Post);
-                ScreenKeyboardEvents.afterKeyPress(screen).register(PackActionsHandler::onKeyPressed$Post);
+            if (screen instanceof PackSelectionScreen packSelectionScreen) {
+                ScreenEvents.afterRender(screen).register((Screen screen1, GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) -> {
+                    PackActionsHandler.onScreen$Render$Post(Screens.getClient(packSelectionScreen), packSelectionScreen, guiGraphics, mouseX, mouseY, partialTick);
+                });
+                ScreenKeyboardEvents.afterKeyPress(screen).register((Screen screen1, int keyCode, int scanCode, int modifiers) -> {
+                    PackActionsHandler.onKeyPressed$Post(Screens.getClient(packSelectionScreen), packSelectionScreen, keyCode, scanCode, modifiers);
+                });
             }
         });
         ClientTickEvents.END_CLIENT_TICK.register(PackActionsHandler::onClientTick$End);