+++ 1.20.4/CHANGELOG.md
@@ -4,5 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v20.4.0-1.20.4] - 2024-01-04
+## [v20.4.0-1.20.4] - 2024-01-05
 - Ported to Minecraft 1.20.4
+### Added
+- Added support for NeoForge
+- Added a new debug key (`T`) for restoring default resource packs list and order
+### Fixed
+- Fixed debug keys partially working on screens other than the resource pack selection screen